reservation

+++ demo/src/main/java/sd_back/demo/repository/member/JpaMemberRepository.java
@@ -5,7 +5,7 @@ import sd_back.demo.domain.Member;
 
 import java.util.Optional;
 
-public interface JpaMemberRepository extends JpaRepository<Member, Integer> {
+public interface JpaMemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByStudentId(Long studentId);
     //기본적인 CRUD 메소드를 바로바로 만들어줌
     //findBy[칼럼명, 첫글자 대문자로]
